app/requirements.txt
fastapi
uvicorn
pymongo
pydantic
slowapi
werkzeug
python-dotenv
python-dateutil
google-generativeai
